feat: skip downloaded files

+++ download.py
@@ -8,7 +8,8 @@ import urllib.request
 URL_BASE = "https://oceans11.lanl.gov/firetec/valley_losAlamos/"
 FILENAME_BASE = "output."
 FILENAME_EXT = ".vts"
-FILE_IDS = range(1000, 75000, 1000)
+EXCLUDED_IDS = (42000,)
+FILE_IDS = filter(lambda id: id not in EXCLUDED_IDS, range(1000, 75000, 1000))
 
 DATA_DIR = pathlib.Path(__file__).parent.resolve() / "data"
 
@@ -16,6 +17,11 @@ DATA_DIR = pathlib.Path(__file__).parent.resolve() / "data"
 def download_file(file_id: int):
     url = f"{URL_BASE}{FILENAME_BASE}{file_id}{FILENAME_EXT}"
     file_path = DATA_DIR / f"{FILENAME_BASE}{file_id}{FILENAME_EXT}"
+
+    if file_path.exists():
+        print(f"Skipping {file_path}...")
+        return
+
     print(f"Downloading {file_path}...")
     file_path.parent.mkdir(parents=True, exist_ok=True)
     with urllib.request.urlopen(url) as response, open(file_path, "wb") as out_file:
